Backend Karriere erstellen

+++ backend/src/main/java/at/htlkaindorf/backend/repositories/PlayerRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface PlayerRepository extends JpaRepository<Player, Long> {
 
     @Query("SELECT p from Player p WHERE p.startClub.club_id = ?1")
-    List<Player> findPlayersByClub(Integer clubId);
+    List<Player> findPlayersByClub(Long clubId);
 
 }
